Fixed a bug using promotion plugin and svnmerge
When using the promotion plugin to perform an upstream integration, the
svn information is not found.  This is due to it not being copied to the
promotion sub-build.  As a quick fix, if the upstream is not found, try
to retrieve it from the root build.  Else it will fail as before.